Move registration I/O from form to parent page
